Moved properties code in FlowsProperties component
